refactorings

+++ Controllers/UserController.js
@@ -2,40 +2,12 @@ const RequestError = require("../Errors/RequestError");
 const UserService = require("../Services/UserService");
 const validator = require("validator");
 const USER_TYPES = require("../Models/User").USER_TYPES;
+const Logger = require("winston").loggers.get("Server");
 
 const createUser = async (req, res, next) => {
 
     const email = req.body.email || "";
     const password = req.body.password || "";
-    const passwordConfirm = req.body.passwordConfirm || "";
-
-
-    if (!email || !password || !passwordConfirm) {
-        return next(new RequestError("email, password and passwordConfirm fields are required!"));
-    }
-
-    if (!validator.isEmail(email))
-        return next(new RequestError("invalid email"));
-
-    if (password.length < 6)
-        return next(new RequestError("password must be at least 6 characters!"));
-
-    if (password !== passwordConfirm)
-        return next(new RequestError("password and passwordConfirm don't match!"));
-
-    try {
-        const result = await UserService.createSimpleUser(email, password);
-        if (!result) return next(new RequestError("User already exists!"));
-        return res.json({success: true});
-    } catch (error) {
-        next(error)
-    }
-}
-const deleteUser = async (req, res, next) => {
-
-    const email = req.body.email || "";
-    const password = req.body.password || "";
-
 
     if (!email || !password) {
         return next(new RequestError("email and password fields are required!"));
@@ -44,8 +16,32 @@ const deleteUser = async (req, res, next) => {
     if (!validator.isEmail(email))
         return next(new RequestError("invalid email"));
 
+    if (password.length < 6)
+        return next(new RequestError("password must be at least 6 characters!"));
 
     try {
+        Logger.info(`User create: ${email}`);
+        const result = await UserService.createSimpleUser(email, password);
+        if (!result) return next(new RequestError("User already exists!"));
+        return res.json({success: true});
+    } catch (error) {
+        next(error)
+    }
+}
+
+const deleteUser = async (req, res, next) => {
+    const email = req.body.email || "";
+    const password = req.body.password || "";
+
+    if (!email || !password) {
+        return next(new RequestError("email and password fields are required!"));
+    }
+
+    if (!validator.isEmail(email))
+        return next(new RequestError("invalid email"));
+
+    try {
+        Logger.info(`User delete: ${email}`);
         const result = await UserService.deleteUser(email, password);
         if (!result) return next(new RequestError("Delete operation failed"));
         return res.json({success: true});
@@ -58,10 +54,9 @@ const changePassword = async (req, res, next) => {
     const email = req.body.email || "";
     const currentPassword = req.body.currentPassword || "";
     const newPassword = req.body.newPassword || "";
-    const newPasswordConfirm = req.body.newPasswordConfirm || "";
 
-    if (!email || !currentPassword || !newPassword || !newPasswordConfirm) {
-        return next(new RequestError("email, currentPassword, newPassword and newPasswordConfirm fields are required!"));
+    if (!email || !currentPassword || !newPassword) {
+        return next(new RequestError("email, currentPassword and newPassword fields are required!"));
     }
 
     if (!validator.isEmail(email))
@@ -70,12 +65,9 @@ const changePassword = async (req, res, next) => {
     if (newPassword.length < 6)
         return next(new RequestError("password must be at least 6 characters!"));
 
-    if (newPassword !== newPasswordConfirm)
-        return next(new RequestError("newPassword and newPasswordConfirm don't match!"));
-
     try {
+        Logger.info(`User change password: ${email}`);
         const result = await UserService.changePassword(email, currentPassword, newPassword);
-
         if (!result) return next(new RequestError("Something went wrong during password change"));
 
         return res.json({success: true});
@@ -100,6 +92,7 @@ const changeUserType = async (req, res, next) => {
         return next(new RequestError("invalid userType, possible userType: LLM, AnnotationService"));
 
     try {
+        Logger.info(`User change user type: ${email}`);
         const result = await UserService.changeUserType(email, userType);
 
         if (!result) return next(new RequestError("Something went wrong during userType change"));
